Synchronize StreamObservers calls.
Here is javadoc of StreamObservers:
Implementations are not required to be thread-safe (but should be thread-compatible ). Separate StreamObservers do not need to be synchronized together; incoming and outgoing directions are independent. Since individual StreamObservers are not thread-safe, if multiple threads will be writing to a StreamObserver concurrently, the application must synchronize calls.